+++ source/clib/tests/cntr/cntr_test_util.c
@@ -99,8 +99,8 @@ static bool binary_search(int *buf, int size, int value) {
 }
 
 int cntr_int_compare(const void* x, const void* y) {
-	int a = (int)x;
-	int b = (int)y;
+	int a = (int)(intptr_t)x;
+	int b = (int)(intptr_t)y;
 
 	if (a < b) return -1;
 	else if (a > b) return 1;
